trying my own version
I have only started on it but am all ready running into walls. I am
going for a website which has each bit of information on a separate page
to put it into words. It won't really be but each bit of information you
will need to scroll a bit and only that section will be on the screen.