changed the layout
I mucked about with the layout. I also realised that the text inside p
tags is not the right font, I have been trying to find the reason but
have not found it yet. I also added a css indent in the title text to
make it look more letter pressed